Refactored TUI application into a proper python project.
