+++ scripts/obc_monitor/__main__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+if __name__ == "__main__":
+    import sys
+    from .main import main
+
+    sys.exit(main())
