+++ scripts/obc_monitor/pyproject.toml
@@ -1,0 +1,3 @@
+[[tool.mypy.overrides]]
+module = "serial.*"
+ignore_missing_imports = true
